requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
httpx==0.27.2
aiohttp==3.13.0
Pillow==10.4.0
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.9
openai>=1.40.0
jinja2>=3.1.4
requests
pandas
openpyxl